requirements.txt
opencv-python
urllib3
charset_normalizer
certifi
scikit-image
onnxruntime
scikit-learn
pyvips
tqdm
requests
slideio
glog
aicsimageio
aicspylibczi>=3.1.1
pycryptodome
h5py
patchify
numba
fastremap
natsort
roifile
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113